Extract css-loader to rules/css
- Pass in arguments as an object
