+++ webpack/webpack.config.js
@@ -15,6 +15,7 @@ const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const PATHS = require('./paths');
 const image = require('./rules/image');
 const javascript = require('./rules/javascript');
+const css = require('./rules/css');
 
 module.exports = (env = '') => {
   const isProd = process.env.NODE_ENV === 'production';
@@ -24,38 +25,6 @@ module.exports = (env = '') => {
   const externals = fs.readdirSync('node_modules')
     .filter(x => ['.bin'].indexOf(x) === -1)
     .reduce((acc, cur) => Object.assign(acc, { [cur]: 'commonjs ' + cur }), {});
-
-  const cssLoader = (production = false, browser = false) => {
-    /*
-     * https://github.com/webpack/css-loader/issues/59 - use css-loader/locals on server-side rendering
-     * :local(.className) used to declare className in the local scope,
-     * local identifiers are exported by css-loader
-     *
-     * modules enable CSS Modules spec https://github.com/css-modules/css-modules
-     *
-     * importLoaders (int): That many loaders after the css-loader are applied to @import-ed resources.
-     *
-     */
-    const queryModules = 'modules';
-    const queryImportLoaders = 'importLoaders=1';
-    const queryLocal = 'localIdentName=[name]__[local]___[hash:base64:5]';
-
-    const queryStrExtractTextPlugin = '?' + queryModules + '&' + queryImportLoaders;
-    const queryStr = queryStrExtractTextPlugin + '&' + queryLocal;
-
-    const serverSide = 'css-loader/locals' + queryStr + '!postcss-loader';
-    const browserSide = 'style-loader!css-loader' + queryStr + '!postcss-loader';
-
-    const myLoader = browser ? browserSide : serverSide;
-
-    const prodBrowserRenderLoader = ExtractTextPlugin.extract({
-      fallbackLoader: 'style-loader',
-      loader: 'css-loader' + queryStrExtractTextPlugin + '!postcss-loader'
-    });
-    const loader = production && browser ? prodBrowserRenderLoader : myLoader;
-    const obj = { test: /\.css$/, include: PATHS.app, loader };
-    return obj;
-  };
 
   const webpackPlugins = (production = false, browser = false) => {
     const bannerOptions = { raw: true, banner: 'require("source-map-support").install();' };
@@ -128,7 +97,10 @@ module.exports = (env = '') => {
             production: isProd,
             browser: false
           }),
-          cssLoader(isProd, false),
+          css({
+            production: isProd,
+            browser: false
+          }),
           image()
       ]
     },
@@ -153,7 +125,10 @@ module.exports = (env = '') => {
             production: isProd,
             browser: true
           }),
-          cssLoader(isProd, true),
+          css({
+            production: isProd,
+            browser: true
+          }),
           image()
         ]
     },
@@ -182,7 +157,10 @@ module.exports = (env = '') => {
             production: isProd,
             browser: true
           }),
-          cssLoader(isProd, true),
+          css({
+            production: isProd,
+            browser: true
+          }),
           image()
         ]
     },
@@ -210,7 +188,10 @@ module.exports = (env = '') => {
             production: isProd,
             browser: false
           }),
-          cssLoader(isProd, false),
+          css({
+            production: isProd,
+            browser: false
+          }),
           image()
         ]
     },
